feat: remain search setting until page refresh
