+++ src/components/app/App.js
@@ -7,6 +7,7 @@ import PageTracking from '../global/PageTracking';
 import Navbar from '../global/Navbar';
 import Footer from '../global/Footer';
 import AppContainer from './AppContainer';
+import SearchSettingContextProvider from '../../utils/context/SearchSettingContextProvider';
 
 // Pages
 const MainPage = lazy(() => import('../../routes/MainPage'));
@@ -20,30 +21,32 @@ const Error404Page = lazy(() => import('../../routes/error/404'));
 const App = () => {
   return (
     <Router>
-      <AppContainer>
-        <Navbar />
+      <SearchSettingContextProvider>
+        <AppContainer>
+          <Navbar />
 
-        {/* MAIN */}
-        <SkeletonTheme color='#ffffff59' highlightColor='#ffffff00'>
-          <PageTracking />
+          {/* MAIN */}
+          <SkeletonTheme color='#ffffff59' highlightColor='#ffffff00'>
+            <PageTracking />
 
-          <Suspense fallback={<div />}>
-            <Switch>
-              <Route exact path='/' component={MainPage} />
-              <Route exact path='/levels' component={LevelListPage} />
-              <Route exact path='/levels/:id' component={LevelPage} />
-              <Route exact path='/mods' component={ModListPage} />
-              <Route exact path='/mods/:id' component={ModPage} />
-              <Route exact path='/ranks' component={RankingPage} />
+            <Suspense fallback={<div />}>
+              <Switch>
+                <Route exact path='/' component={MainPage} />
+                <Route exact path='/levels' component={LevelListPage} />
+                <Route exact path='/levels/:id' component={LevelPage} />
+                <Route exact path='/mods' component={ModListPage} />
+                <Route exact path='/mods/:id' component={ModPage} />
+                <Route exact path='/ranks' component={RankingPage} />
 
-              <Route component={Error404Page} />
-            </Switch>
-          </Suspense>
-        </SkeletonTheme>
+                <Route component={Error404Page} />
+              </Switch>
+            </Suspense>
+          </SkeletonTheme>
 
-        {/* FOOTER */}
-        <Footer />
-      </AppContainer>
+          {/* FOOTER */}
+          <Footer />
+        </AppContainer>
+      </SearchSettingContextProvider>
     </Router>
   );
 };
